backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
aiofiles>=24.1.0,<25.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.7.0,<3.0.0

# Agent dependencies (pulled from Process-Re-Imagination-Agent)
langgraph>=0.2.60,<0.4.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.4.0
mermaid-python==0.1
python-magic>=0.4.27,<1.0.0
pytesseract>=0.3.10,<1.0.0
Pillow>=10.0.0,<11.0.0
pdf2image>=1.16.0,<2.0.0
pypdf>=5.0.0,<6.0.0
python-pptx>=1.0.2,<2.0.0
lxml>=5.2.0,<6.0.0
requests>=2.31.0,<3.0.0
